requirements.txt
numpy<2
matplotlib
scipy
pyyaml
pillow
tqdm
kornia==0.7.3
gdown
hjson
ujson
timm
evo
trimesh
munch
plyfile
glfw
PyOpenGL
PyGLM
rich
ruff
configargparse
numba
scikit-learn
scikit-image
rerun-sdk
opencv-python
